LayoutTests/http/tests/security/cross-frame-access-port-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://127.0.0.1:8080/security/resources/cross-frame-iframe.html from frame with URL http://127.0.0.1:8000/security/cross-frame-access-port.html. Domains, protocols and ports must match.


PASS: Cross frame access to different port was denied!


--------
Frame: 'aFrame'
--------
Inner iframe.
